_PROJECTS/AURENIX AGENCY/aurenix-monorepo/legacy/services/temporal-worker/requirements.txt
temporalio
pydantic
pyyaml
python-dotenv
supabase
